Container for canvas wrapper
