+++ cvat-ui/src/containers/annotation-page/standard-workspace/canvas-wrapper.tsx
@@ -1,0 +1,59 @@
+import React from 'react';
+import { connect } from 'react-redux';
+
+import CanvasWrapperComponent from '../../../components/annotation-page/standard-workspace/canvas-wrapper';
+
+import {
+    confirmCanvasReady,
+} from '../../../actions/annotation-actions';
+import {
+    CombinedState,
+} from '../../../reducers/interfaces';
+
+import { Canvas } from '../../../canvas';
+
+interface StateToProps {
+    canvasInstance: Canvas;
+    jobInstance: any;
+    annotations: any[];
+    frameData: any;
+}
+
+interface DispatchToProps {
+    onSetupCanvas(): void;
+}
+
+function mapStateToProps(state: CombinedState): StateToProps {
+    const {
+        canvasInstance,
+        jobInstance,
+        frameData,
+        annotations,
+    } = state.annotation;
+
+    return {
+        canvasInstance,
+        jobInstance,
+        frameData,
+        annotations,
+    };
+}
+
+function mapDispatchToProps(dispatch: any): DispatchToProps {
+    return {
+        onSetupCanvas(): void {
+            dispatch(confirmCanvasReady());
+        },
+    };
+}
+
+function CanvasWrapperContainer(props: StateToProps & DispatchToProps): JSX.Element {
+    return (
+        <CanvasWrapperComponent {...props} />
+    );
+}
+
+export default connect(
+    mapStateToProps,
+    mapDispatchToProps,
+)(CanvasWrapperContainer);
